Reduce complexity
