+++ src/client/api/index.ts
@@ -2,3 +2,4 @@ export { fetchLeaderboardData } from './requests';
 export { useGetTeam } from './hooks/use-get-team';
 export { useLoadTeams } from './hooks/use-load-teams';
 export { useUpdateTeam } from './hooks/use-update-team';
+export { prizeMoneyState, teamsState } from './state/atoms';
