+++ src/client/leaderboard/components/body-container.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useLoadTeams } from 'src/client/api';
 import { TeamType } from '../../../types';
-import { useTeamsMoney } from '../hooks/use-teams-money';
+import { useAddPrizeMoney } from '../hooks/use-add-prize-money';
 import { TableBody } from './table-body';
 
 const prizeMoneySortKey = (a: TeamType, b: TeamType) =>
@@ -10,8 +10,8 @@ const prizeMoneySortKey = (a: TeamType, b: TeamType) =>
 export const BodyContainer = () => {
   // Needs to be done here to pull in any team changes
   useLoadTeams();
-  const prizeMoneyData = useTeamsMoney();
-  const sortedData = prizeMoneyData.slice().sort(prizeMoneySortKey);
+  const teamsWithPrizeMoney = useAddPrizeMoney();
+  const sortedData = teamsWithPrizeMoney.slice().sort(prizeMoneySortKey);
 
   return <TableBody tableData={sortedData} />;
 };
