+++ src/client/util/prize-money.ts
@@ -43,35 +43,3 @@ const splitMoneyOnTie = (position: number, noPlayersTied: number) => {
   // Split the money between the players
   return Math.round(positionPrizeMoney / noPlayersTied);
 };
-
-export const allTeamsMoney = (
-  golfers: Golfers,
-  rankingsWithPrizeMoney: GolferMoneyRankings,
-  teams: TeamType[]
-) => {
-  // Loop through entrants
-  // Loop through player IDs and get the players position
-  // Use the position to get their prize money
-
-  return teams.map((team) => teamMoney(team, golfers, rankingsWithPrizeMoney));
-};
-
-const teamMoney = (
-  team: TeamType,
-  golfers: Golfers,
-  rankingsWithPrizeMoney: GolferMoneyRankings
-) => {
-  const teamPrizeMoney = team.golfer_ids.reduce((accum, id) => {
-    try {
-      const position = golfers[id].position;
-      return accum + rankingsWithPrizeMoney[position].prizeMoney;
-    } catch (err) {
-      throw new Error(`No ID for ${id}`);
-    }
-  }, 0);
-
-  return {
-    ...team,
-    prizeMoney: teamPrizeMoney,
-  };
-};
